more firecracker fixes
