+++ pkg/virtualizers/firecracker/virtualizer_test.go
@@ -2,8 +2,6 @@ package firecracker
 
 import (
 	"testing"
-
-	logger "github.com/vorteil/vorteil/pkg/virtualizers/logging"
 )
 
 var codeBlockToLookIP = `
@@ -34,16 +32,5 @@ func TestType(t *testing.T) {
 	typeSt := v.Type()
 	if typeSt != "vmware" {
 		t.Errorf("expected %s but got %s", "vmware", typeSt)
-	}
-}
-func TestLoggerAndSerial(t *testing.T) {
-	v := &Virtualizer{
-		serialLogger: logger.NewLogger(2048),
-	}
-
-	seriall := v.Serial()
-
-	if seriall == nil {
-		t.Errorf("unable to get loggers from virtualizer")
 	}
 }
